final project submit

+++ event/views.py
@@ -4,8 +4,12 @@ import calendar
 from calendar import HTMLCalendar
 from django.http import HttpResponseRedirect , HttpResponse
 from .models import Event , Venue
-from .forms import VenueForm , EventForm
+## import user model 
+from django.contrib.auth.models import User
+
+from .forms import VenueForm , EventForm ,EventFormAdmin
 import csv
+from django.contrib import messages
 
 # PDF stuff 
 from django.http import FileResponse
@@ -16,6 +20,49 @@ from reportlab.lib.pagesizes import LETTER
 
 #pagination stuff 
 from  django.core.paginator import Paginator
+
+
+
+
+def admin_approval(request):
+    events_count= Event.objects.all().count()
+    venue_count= Venue.objects.all().count()
+    users_count = User.objects.all().count()
+    events= Event.objects.all()
+    if request.user.is_superuser:
+        if request.method =='POST':
+        
+            id_list = request.POST.getlist('boxes')
+            events.update(approved=False)
+            for x in id_list:
+                Event.objects.filter(pk=int(x)).update(approved = True)
+
+            return redirect('events')
+        else:
+            return render(request , 'event/admin_approval.html' , {
+                'events':events,
+                'events_count':events_count,
+                'venue_count':venue_count,
+                'users_count':users_count,
+                })
+    else : 
+        messages.error(request , (' you are not authorized '))
+        return redirect('events')
+    
+    return render(request , 'event/admin_approval.html' , {'events':events})
+
+def myEvent(request):
+    if request.user.is_authenticated:
+        me = request.user.id
+        events = Event.objects.filter(manager= me)  
+        
+        #print(events)  
+        return render(request, 'event/myEvent.html', {"events": events})
+
+    else:
+        messages.error(request, 'You are not authorized')
+        return redirect('home')
+
 
 
 def download_venues(request):
@@ -83,8 +130,13 @@ def venue_txt(request):
 
 def delete_event(request , event_id):
     event = Event.objects.get(pk= event_id)
-    event.delete()
-    return redirect('/events')
+    if request.user == event.manager:
+        messages.success(request , 'it has been deleted succesfully')
+        event.delete()
+        return redirect('/events')
+    else : 
+        messages.error(request , 'user is not authorized to delete this .')
+        return redirect('/events')
 
 def delete_venue(request , venue_id):
     event = Venue.objects.get(pk= venue_id)
@@ -95,21 +147,33 @@ def delete_venue(request , venue_id):
 def event_Form(request):
     submitted= False
     if request.method == 'POST':
-        form=EventForm(request.POST)
-        if form.is_valid():
-            form.save()
-            return HttpResponseRedirect('/events?submitted')
-    else :
-         form = EventForm()
-         if 'submitted' in request.GET :
-             submitted= True
+        if request.user.is_superuser:   # this mean id of superuser in django , if iwant someone else i can make it like request.user.id == 2 for ex.
+            form=EventFormAdmin(request.POST)
+            if form.is_valid():
+                form.save()
+                return HttpResponseRedirect('/events?submitted')
+        else:
+            form=EventForm(request.POST)
+            if form.is_valid():
+                event = form.save(commit=False)
+                event.manager = request.user
+                event.save()
+                return HttpResponseRedirect('/events?submitted')
+    else:
+        if request.user.is_superuser:
+            form = EventFormAdmin
+        else :   
+            form = EventForm
+
+            if 'submitted' in request.GET:
+                submitted= True
    
     return render(request , 'event/Event_form.html' , {'form':form , 'submitted': submitted})
 
 
 def update_venue(request ,venue_id ):
     venue = Venue.objects.get(pk= venue_id)
-    form=VenueForm(request.POST or None , instance=venue)
+    form=VenueForm(request.POST or None ,request.FILES or None , instance=venue)
     if form.is_valid():
         form.save()
         return redirect("venues")
@@ -117,7 +181,11 @@ def update_venue(request ,venue_id ):
 
 def update_event(request ,event_id ):
     event = Event.objects.get(pk= event_id)
-    form=EventForm(request.POST or None , instance=event)
+    if request.user.is_superuser:
+        form=EventFormAdmin(request.POST or None , instance=event)
+    else: 
+        form=EventFormAdmin(request.POST or None , instance=event)
+
     if form.is_valid():
         form.save()
         return redirect("venues")
@@ -138,7 +206,8 @@ def search_venue(request):
 
 def show_venue(request , venue_id):
     venue = Venue.objects.get(pk= venue_id)
-    return render(request, 'event/show_venue.html', {'venue': venue})
+    venueOwner= User.objects.get(pk=venue.owner)
+    return render(request, 'event/show_venue.html', {'venue': venue , 'venueOwner': venueOwner})
 
 
 
@@ -158,21 +227,23 @@ def venue_l(request):
 def Venue_Form(request):
     submitted= False
     if request.method == 'POST':
-        form=VenueForm(request.POST)
+        form=VenueForm(request.POST , request.FILES)
         if form.is_valid():
+            venue = form.save(commit=False)
+            venue.owner = request.user.id
             form.save()
             return HttpResponseRedirect('/venues?submitted')
     else :
-         form = VenueForm()
-         if 'submitted' in request.GET :
-             submitted= True
-   
+        form = VenueForm()
+        if 'submitted' in request.GET :
+            submitted= True
+
     return render(request , 'event/Venue_form.html' , {'form':form , 'submitted': submitted})
 
 def event_l(request):
     #event_list = Event.objects.all().order_by('name')
 
-    p = Paginator(Event.objects.all() , 2)
+    p = Paginator(Event.objects.all() , 8)
     page = request.GET.get('page')
     events = p.get_page(page)
 
@@ -203,12 +274,15 @@ def homepage(request, year=None, month=None):
     # Create the calendar
     cal = HTMLCalendar().formatmonth(year, month_number)
 
+    event_list = Event.objects.filter(date__year = datetime.now().year , date__month = datetime.now().month)
+
     context = {
         'name': name,
         'month_number': month_number,
         'cal': cal,
         'year': year,
         'month': month,
+        'event_list':event_list
     }
     return render(request, 'event/home.html', context)
